Split layout of site and block preview
