+++ site/src/app/block-preview/page/page.tsx
@@ -1,10 +1,12 @@
 "use client";
 import { BlockPreviewProvider, IFrameBridgeProvider, useBlockPreviewFetch, useIFrameBridge } from "@comet/cms-site";
+import { ContentScope } from "@src/../../site-configs.d";
 import { PageContentBlockData } from "@src/blocks.generated";
 import { PageContentBlock } from "@src/documents/pages/blocks/PageContentBlock";
 import { graphQLApiUrl } from "@src/util/graphQLClient";
 import { recursivelyLoadBlockData } from "@src/util/recursivelyLoadBlockData";
 import { useEffect, useState } from "react";
+import { IntlProvider } from "react-intl";
 
 const PreviewPage: React.FunctionComponent = () => {
     const iFrameBridge = useIFrameBridge();
@@ -31,11 +33,22 @@ const PreviewPage: React.FunctionComponent = () => {
 
     return <div>{blockData && <PageContentBlock data={blockData} />}</div>;
 };
+
+function PreviewPageWrapper() {
+    const { contentScope } = useIFrameBridge() as { contentScope?: ContentScope };
+    if (!contentScope) return null;
+    return (
+        <IntlProvider defaultLocale={contentScope.language} locale={contentScope.language}>
+            <PreviewPage />
+        </IntlProvider>
+    );
+}
+
 const IFrameBridgePreviewPage = () => {
     return (
         <IFrameBridgeProvider>
             <BlockPreviewProvider>
-                <PreviewPage />
+                <PreviewPageWrapper />
             </BlockPreviewProvider>
         </IFrameBridgeProvider>
     );
